Semana 7

+++ src/main/java/edu/co/icesi/introspringboot/repository/ProfessorRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ProfessorRepository extends CrudRepository<Professor, Long> {
+public interface ProfessorRepository extends CrudRepository<Professor, Integer> {
 
     List<Professor> findByNameContainingIgnoreCase(String segment);
